Renamed module to pyopenhardwaremonitor
