+++ .github/workflows/code-checker.yml
@@ -14,7 +14,7 @@ jobs:
           - "3.12"
           - "3.13"
     env:
-      SRC_FOLDER: openhardwaremonitor
+      SRC_FOLDER: pyopenhardwaremonitor
     steps:
       - uses: actions/checkout@v2
       - name: Set up Python ${{ matrix.python-version }}
