--- openhardwaremonitor/const.py
@@ -1,3 +1,0 @@
-"""Constants used by pyOpenHardwareMonitor"""
-
-__version__ = "0.1.3"
