+++ pyopenhardwaremonitor/__init__.py
