+++ pyopenhardwaremonitor/api.py
