+++ pyopenhardwaremonitor/const.py
@@ -1,0 +1,3 @@
+"""Constants used by Open Hardware Monitor"""
+
+__version__ = "0.1.3"
