+++ setup.cfg
@@ -1,2 +1,2 @@
 [metadata]
-name = pyOpenHardwareMonitor
+name = pyopenhardwaremonitor
